requirements_phone.txt
# Phone Voice Participation System Requirements

# Core async/websocket support
websockets>=11.0.0
asyncio-mqtt>=0.16.0

# Audio processing
numpy>=1.21.0
scipy>=1.7.0

# WebSocket and HTTP server
websockets>=11.0.0

# Deepgram STT integration
deepgram-sdk>=3.0.0

# QR code generation
qrcode[pil]>=7.4.0

# Existing voice system dependencies
openai>=1.0.0
anthropic>=0.25.0
python-dotenv>=1.0.0
PyYAML>=6.0.0
pyaudio>=0.2.11
elevenlabs>=0.2.0

# Audio format conversion (requires ffmpeg installed separately)
# Install ffmpeg: brew install ffmpeg (macOS) or apt install ffmpeg (Ubuntu)

# Optional: For better audio processing
librosa>=0.10.0  # Advanced audio analysis
soundfile>=0.12.0  # Audio file I/O 